# doc/api.yaml
openapi: 3.0.1
info:
  title: WASAPhoto APIs
  version: 1.0.3
  description: >
    **"Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto!"**
    
    WASAPhoto is web application based on RESTfull architecture to allows upload your photos directly from your PC, and they will be visible to everyone following you.
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  
  contact:
    name: Andrea Pizzi
    url: https://github.com/pizzi-andrea/WASAPhoto
    email: pizzi.1995517@studenti.uniroma1.it
servers:
  - url: http://localhost:3000
paths:
...
  /users/{uid}/followers/{followerId}:
    summary: a follower
    description: >
      this resource rappresent follower
    parameters:
      - $ref: "#/components/parameters/Uid"
      - $ref: "#/components/parameters/FollowerId"
    delete:
      tags:
        - user
        - follower
      summary: unfollow user
      description: >
        gived uid and *followedId* then remove follower *followerId* from user followers
      security:
        - BearerAuth: []
        
      operationId: unfollowUser
      
      responses:
        "204":
          description: >
            success , unfollow the user
        "404":
          description: >
            Not found, just no follow user or user Id not exist
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
    
    put:
      tags:
        - user
        - follower
      summary:  follow user
      description: >
        give user  and added to  follower of user
      
      
      requestBody:
        description: >
          information on followers
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      
      security:
        - BearerAuth: []
      
      operationId: followUser
      
      responses:
        "201": 
          description: >
            added new follower
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "204":
          description: >
            Empty response, just follow the user
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{uid}/banned/{bannedId}:
    summary: users banned
    description: >
      this resource describe banned user
    
    parameters:
      - $ref: "#/components/parameters/Uid"
      - $ref: "#/components/parameters/BannedId"
    
    put:
      tags:
        - user
        - ban
      summary: ban user identificated by *uid* 
      description: >
        taken uid of the user who wants to ban and uid of the user to be banned, 
        bans the user(BannedID)
              
      security:
        - BearerAuth: []
      
      operationId: banUser
      
      responses:
        "201":
          description: >
            success, banned user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "204":
          description: >
            empty response, user just banned
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
    
    delete:
      tags:
        - user
        - ban
      summary: unban user
      description: >
        taken uid of the user who wants unbband id of user banned and delete last one
      security:
        - BearerAuth: []
        
      operationId: unbanUser
      
      responses:
        "204":
          description: >
            success, unbanned user
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{uid}/myPhotos/{photoId}/: #
    summary: photo updated by user
    description: >
      this resource rappresent the streams of photos update by user in time
  
    parameters:
      - $ref: "#/components/parameters/Uid"
      - $ref: "#/components/parameters/PhotoId"
          
    delete:
      tags:
        - photo
      summary: delete photo updated
      description: >
        provide uid and photoID and delete the photo corresponding to photoID
    
      operationId: deletePhoto
      security:
        - BearerAuth: []
        
      responses:
        "204":
          description: >
            success, deleted photo
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
    
    get:
      tags:
        - photo
      security:
        - BearerAuth: []
      summary: get photo
      description: >
        give *uid* and photo id and get photo with photo id gived
      operationId: getPhoto
      
      responses:
        "200": 
          description: >
            success, photo founded
          
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Photo"
          
        "404":
          $ref: "#/components/responses/UserNotFound"
      
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        
        "204":
          description: >
            empty body, found nothing
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{uid}/myPhotos/{photoId}/comments/:
    
    parameters:
      - $ref: "#/components/parameters/Uid"
      - $ref: "#/components/parameters/PhotoId"
    
    post:
      tags:
        - comment
        - photo
      summary: add comment a photo
      description: >
        give user id and photo id if user is owner of photo add a comment
        stored in the body of request
      
      
      
      requestBody:
        description: > 
          comment to add
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/Comment"
      
      security:
        - BearerAuth: []
      operationId: commentPhoto
      
      responses:
        "201":
          description: >
            success, new comment added to photo
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Comment"
          
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
    
    get:
      security:
        - BearerAuth: []
        
      tags:
        - photo
        - comment
      summary: get comments on photo
      description: >
        give photo id and user id of owner photo then get all comments on photo
      
      parameters:
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/UsernameFilter"
        - name: sortByData
          in: query
          required: false
          schema:
            description: >
              flag for sorting comments on photo by update data
            type: boolean
            default: true
          description: >
            sort comments by data creation
        
      
      operationId: getComments
      responses:
        "200":
          description: >
            success, get all comments on photo
          
          content:
            application/json:
              schema:
                description: >
                  list of comments on photo
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
                minItems: 1
                maxItems: 500

        "404":
          $ref: "#/components/responses/UserNotFound"
        "204":
          description: >
            empty body, found nothing
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /session/:

    post:
      tags: 
        - login
      summary: Logs in the user
      description: >
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      
      operationId: doLogin
     
      requestBody:
        description: >
          User details
        content:
          application/json:
            schema:
              description: >
                username
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/Username"
        required: true
      responses:
        "201":
          description: >
            User log-in action successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
components:
  parameters:
    QueryLimit:
      description: >
        security parameter to limit the number of items in a response and avoid DOS attacks
      name: limit
      in: query
      required: false
      schema:
        description: >
          format limit values
        type: integer
        format: int
        minimum: 1
        example: 50
    UsernameFilter:
      description: >
        research in list of users by username 
      name: username
      in: query
      required: false
      schema:
        type: string
        pattern: '^.*?$'
        minLength: 1
        maxLength: 16
        example: "marioross"
    Uid:
      name: uid
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
      description: >
        user id
    PhotoId:
      name: photoId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
      description: >
        photo id
...
    FollowerId:
      name: followerId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
      description: >
        user to unfollow
    BannedId:
      name: bannedId
      in: path
      required: true
      description: >
        uid user to ban
      schema:
        $ref: "#/components/schemas/Id"
  responses:
    UnauthorizedError:
      description: >
        Access token is missing or invalid 
    UnauthorizedToken:
      description: >
        the submitted token does not have the necessary permissions to access this resource
    ServerError:
      description: >
        server internal error, for more information see log file
    UserNotFound:
      description: >
        Not found, user not found
    BadRequest:
      description: >
       Bad request, formating error 
  securitySchemes:
    
    BearerAuth:
      description: >
        basic safety mechanism based on "bearer" mechanism. APIs that require a minimum level of security can adopt this method
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      description: >
        this object rappresent a user
      type: object
      properties:
        uid:
          $ref: "#/components/schemas/Id"
        username: 
          $ref: "#/components/schemas/Username"
      required:
        - uid
        - username
    Token: 
      description: >
        This object rappresent a token identification. The token will be used by users to authenticate to the system.
        The token is composed from : 
          - value,  secred token identifier  
      type: object
      properties:
        value:
          $ref: "#/components/schemas/Id"
    Username:
      description: >
        username of a user
      type: string
      pattern: '^.*?$'
      minLength: 3
      maxLength: 16
      example: "marioross"
    Id:
      description: >
        Identificator at 64-bit
      type: integer
      format: int64 
      readOnly: true
      example: 4570 
...
    Photo:
      description: >
        this object rappresent a photo
      type: object
      properties:
        photoId:
          $ref: "#/components/schemas/Id"
        timeUpdate:
          $ref: "#/components/schemas/TimeStamp"

        imageData:
          description: >
            data
          type: string
          format: binary
          minLength: 0
          maxLength: 5000000
          

        
        descriptionImg:
          description: >
            image description
          type: string
          example: "a good photo!"
          minLength: 1
          maxLength: 250
          pattern: "^.*?$"
      required:
        - imageData
        - timeUpdate
...
    TimeStamp:
      description: >
        this components describe timestamp value
        conform to [RFC3339](https://datatracker.ietf.org/doc/html/rfc3339#section-5.6) specification
      type: string
      format: data-time
      pattern: '^[1-9]\d{3}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$'
      example: "2017-07-21T17:32:28Z"
      minLength: 20
      maxLength: 21
      readOnly: true
    Comment:
      description: > 
        this object rappresent a comment on a photo.
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        commentId:
          $ref: "#/components/schemas/Id"
        text:
          description: >
            comment text encoded in UNICODE format
          type: string
          minLength: 1
          maxLength: 250
          pattern: "^.*$"
          example: 
            "😀 i like you photo! 😀"
        timeStamp:
          $ref: "#/components/schemas/TimeStamp"
      required:
        - user
        - text
        - timeStamp
        - commentId
...